update github workflow and pages for static export
